+++ inst/sql/sql_server/createTargetOutcomeCombinations.sql
@@ -1,13 +1,15 @@
 --need to know indication/target/outcome tuples
-drop table if exists #targets_agg_all;
-select * into #targets_agg_all
+
+-- all targets
+drop table if exists #targets_all;
+select * into #targets_all
 from @target_database_schema.@target_table
 where cohort_definition_id in
 (@target_ids);
 
 -- first T with > minPrioObs
-drop table if exists #targets_agg;
-select * into #targets_agg
+drop table if exists #targets_inclusions;
+select * into #targets_inclusions
 from
 (select *,
 row_number() over(partition by subject_id, cohort_definition_id order by cohort_start_date asc) as rn
@@ -22,135 +24,32 @@ and temp_t.cohort_start_date <= op.observation_period_end_date
 where temp_t.rn = 1
 and datediff(day, op.observation_period_start_date, temp_t.cohort_start_date) >= @min_prior_observation;
 
-drop table if exists #outcomes_agg;
-select * into #outcomes_agg
+-- all outcomes
+drop table if exists #outcomes_all;
+select * into #outcomes_all
 from @outcome_database_schema.@outcome_table
 where cohort_definition_id in
 (@outcome_ids);
 
--- first outcomes
-drop table if exists #outcomes_agg_first;
-select * into #outcomes_agg_first
+-- first outcomes in washout days and min prior obs
+drop table if exists #outcomes_washout;
+select o.* into #outcomes_washout
 from (select *,
-row_number() over(partition by subject_id, cohort_definition_id order by cohort_start_date asc) as rn
-from #outcomes_agg
+ISNULL(datediff(day, LAG(cohort_start_date) OVER(partition by subject_id, cohort_definition_id order by cohort_start_date asc), cohort_start_date ), 100000) as time_between
+from #outcomes_all
 ) as o
-where o.rn = 1
-;
-
--- create all the cohort details
-drop table if exists #cohort_details;
-
-select *,
-ROW_NUMBER() OVER (ORDER BY cohort_type, target_cohort_id, outcome_cohort_id) as cohort_definition_id
-into #cohort_details
-from
-
-(
-select distinct
-t.cohort_definition_id as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'TnO' as cohort_type
-from
-(select distinct cohort_definition_id from #targets_agg) as t
-CROSS JOIN
-(select distinct cohort_definition_id from #outcomes_agg) as o
-
-union
-
-select distinct
-t.cohort_definition_id as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'OnT' as cohort_type
-from
-(select distinct cohort_definition_id from #targets_agg) as t
-CROSS JOIN
-(select distinct cohort_definition_id from #outcomes_agg) as o
-
-union
-
---select distinct
---t.cohort_definition_id as target_cohort_id,
---o.cohort_definition_id as outcome_cohort_id,
---'TnOc' as cohort_type
---from
---(select distinct cohort_definition_id from #targets_agg) as t
---CROSS JOIN
---(select distinct cohort_definition_id from #outcomes_agg) as o
-
---union
-
-select distinct
-t.cohort_definition_id as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'TnfirstO' as cohort_type
-from
-(select distinct cohort_definition_id from #targets_agg) as t
-CROSS JOIN
-(select distinct cohort_definition_id from #outcomes_agg) as o
-
-union
-
-select distinct
-t.cohort_definition_id as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'firstOnT' as cohort_type
-from
-(select distinct cohort_definition_id from #targets_agg) as t
-CROSS JOIN
-(select distinct cohort_definition_id from #outcomes_agg) as o
-
-union
-
---select distinct
---t.cohort_definition_id as target_cohort_id,
---o.cohort_definition_id as outcome_cohort_id,
---'TnfirstOc' as cohort_type
---from
---(select distinct cohort_definition_id from #targets_agg) as t
---CROSS JOIN
---(select distinct cohort_definition_id from #outcomes_agg) as o
-
---union
-
-select distinct
-t.cohort_definition_id as target_cohort_id,
-0 as outcome_cohort_id,
-'T' as cohort_type
-from (select distinct cohort_definition_id from #targets_agg) as t
-
-union
-
-select distinct
-t.cohort_definition_id as target_cohort_id,
-0 as outcome_cohort_id,
-'allT' as cohort_type
-from (select distinct cohort_definition_id from #targets_agg) as t
-
-union
-
-select distinct
-0 as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'O' as cohort_type
-from (select distinct cohort_definition_id from #outcomes_agg) as o
-
-union
-
-select distinct
-0 as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id,
-'firstO' as cohort_type
-from (select distinct cohort_definition_id from #outcomes_agg) as o
-
-
-) temp;
+inner join @cdm_database_schema.observation_period op
+on op.person_id = o.subject_id
+and o.cohort_start_date >= op.observation_period_start_date
+and o.cohort_start_date <= op.observation_period_end_date
+where o.time_between >= @outcome_washout_days
+and datediff(day, op.observation_period_start_date, o.cohort_start_date) >= @min_prior_observation;
 
 
 -- 1) get all the people with the outcome in TAR
-drop table if exists #target_with_outcome;
+drop table if exists #target_outcome_tar;
 
--- TnO
+-- cases
 select
 t.subject_id,
 t.cohort_start_date,
@@ -159,28 +58,8 @@ o.cohort_start_date as outcome_start_date,
 o.cohort_end_date as outcome_end_date,
 t.cohort_definition_id as target_cohort_id,
 o.cohort_definition_id as outcome_cohort_id
-into #target_with_outcome
-from #targets_agg t inner join #outcomes_agg o
-on t.subject_id = o.subject_id
-where
--- outcome starts before TAR end
-o.cohort_start_date <= dateadd(day, @tar_end, t.@tar_end_anchor)
-and
--- outcome starts (ends?) after TAR start
-o.cohort_start_date >= dateadd(day, @tar_start, t.@tar_start_anchor);
-
--- TnfirstO
-drop table if exists #target_outcome_f;
-select
-t.subject_id,
-t.cohort_start_date,
-t.cohort_end_date,
-o.cohort_start_date as outcome_start_date,
-o.cohort_end_date as outcome_end_date,
-t.cohort_definition_id as target_cohort_id,
-o.cohort_definition_id as outcome_cohort_id
-into #target_outcome_f
-from #targets_agg t inner join #outcomes_agg_first o
+into #target_outcome_tar
+from #targets_inclusions t inner join #outcomes_washout o
 on t.subject_id = o.subject_id
 where
 -- outcome starts before TAR end
@@ -190,55 +69,39 @@ and
 o.cohort_start_date >= dateadd(day, @tar_start, t.@tar_start_anchor);
 
 
--- 2) get all the people without the outcome in TAR
---drop table if exists #target_nooutcome;
---select
---t.subject_id,
---t.cohort_start_date,
---t.cohort_end_date,
---t.cohort_definition_id as target_cohort_id,
---o.cohort_definition_id as outcome_cohort_id
---into #target_nooutcome
---from #targets_agg t
---CROSS JOIN
---( select distinct cohort_definition_id from #outcomes_agg) o
---left outer join #target_with_outcome two
---on t.cohort_definition_id = two.target_cohort_id
---and t.subject_id = two.subject_id
---and o.cohort_definition_id = two.outcome_cohort_id
---where two.subject_id IS NULL;
+-- 2) get all the people with the outcome during washout
+drop table if exists #target_outcome_prior;
 
---drop table if exists #target_noout_f;
---select
---t.subject_id,
---t.cohort_start_date,
---t.cohort_end_date,
---t.cohort_definition_id as target_cohort_id,
---o.cohort_definition_id as outcome_cohort_id
---into #target_noout_f
---from #targets_agg t
---CROSS JOIN
---( select distinct cohort_definition_id from #outcomes_agg) o
---left outer join #target_outcome_f two
---on t.cohort_definition_id = two.target_cohort_id
---and t.subject_id = two.subject_id
---and o.cohort_definition_id = two.outcome_cohort_id
---where two.subject_id IS NULL;
+-- people with outcome prior
+select
+t.subject_id,
+t.cohort_start_date,
+t.cohort_end_date,
+t.cohort_definition_id as target_cohort_id,
+o.cohort_definition_id as outcome_cohort_id
+into #target_outcome_prior
+from #targets_inclusions t inner join #outcomes_washout o
+on t.subject_id = o.subject_id
+where
+-- outcome starts and ends within prior_outcome_washout_days
+o.cohort_start_date >= dateadd(day, -@outcome_washout_days, t.cohort_start_date)
+and
+o.cohort_start_date <= dateadd(day, -1, t.cohort_start_date);
 
--- Final: select into #agg_cohorts
 
-select * into #agg_cohorts
+drop table if exists #agg_cohorts_before;
+select * into #agg_cohorts_before
 
 from
 (
--- T with O indexed at T
+-- cases indexed at T
 
 select
 tno.subject_id,
 tno.cohort_start_date,
 tno.cohort_end_date,
 cd.cohort_definition_id
-from #target_with_outcome tno
+from #target_outcome_tar tno
 INNER JOIN #cohort_details cd
 on cd.target_cohort_id = tno.target_cohort_id
 and cd.outcome_cohort_id = tno.outcome_cohort_id
@@ -246,131 +109,124 @@ and cd.cohort_type = 'TnO'
 
 union
 
--- T with first O indexed at T
+-- targets with O prior during washout
 
 select
 tno.subject_id,
 tno.cohort_start_date,
 tno.cohort_end_date,
 cd.cohort_definition_id
-from #target_outcome_f tno
+from #target_outcome_prior tno
 INNER JOIN #cohort_details cd
 on cd.target_cohort_id = tno.target_cohort_id
 and cd.outcome_cohort_id = tno.outcome_cohort_id
-and cd.cohort_type = 'TnfirstO'
+and cd.cohort_type = 'TnOprior'
 
 union
-
--- T with O indexed at O
-
-select
-tno.subject_id,
-tno.outcome_start_date as cohort_start_date,
-tno.outcome_end_date as cohort_end_date,
-cd.cohort_definition_id
-from #target_with_outcome tno
-INNER JOIN #cohort_details cd
-on cd.target_cohort_id = tno.target_cohort_id
-and cd.outcome_cohort_id = tno.outcome_cohort_id
-and cd.cohort_type = 'OnT'
-
-union
-
--- T with first O indexed at O
-
-select
-tno.subject_id,
-tno.outcome_start_date as cohort_start_date,
-tno.outcome_end_date as cohort_end_date,
-cd.cohort_definition_id
-from #target_outcome_f tno
-INNER JOIN #cohort_details cd
-on cd.target_cohort_id = tno.target_cohort_id
-and cd.outcome_cohort_id = tno.outcome_cohort_id
-and cd.cohort_type = 'firstOnT'
-
-
-union
-
--- T without O
-
---select
---tnoc.subject_id,
---tnoc.cohort_start_date,
---tnoc.cohort_end_date,
---cd.cohort_definition_id
---from #target_nooutcome tnoc
---INNER JOIN #cohort_details cd
---on cd.target_cohort_id = tnoc.target_cohort_id
---and cd.outcome_cohort_id = tnoc.outcome_cohort_id
---and cd.cohort_type = 'TnOc'
-
---union
-
--- T without first O
-
---select
---tnoc.subject_id,
---tnoc.cohort_start_date,
---tnoc.cohort_end_date,
---cd.cohort_definition_id
---from #target_noout_f tnoc
---INNER JOIN #cohort_details cd
---on cd.target_cohort_id = tnoc.target_cohort_id
---and cd.outcome_cohort_id = tnoc.outcome_cohort_id
---and cd.cohort_type = 'TnfirstOc'
-
---union
-
--- Ts and Os
 
 select distinct * from (
+
+-- targets with restrictions
 
 select
 t.subject_id,
 t.cohort_start_date,
 t.cohort_end_date,
 cd.cohort_definition_id
-from #targets_agg as t
+from #targets_inclusions as t
 INNER JOIN #cohort_details cd
 on cd.target_cohort_id = t.cohort_definition_id
 and cd.cohort_type = 'T'
 
 union
 
+-- outcomes with restrictions
+select
+o.subject_id,
+o.cohort_start_date,
+o.cohort_end_date,
+cd.cohort_definition_id
+from #outcomes_washout as o
+INNER JOIN #cohort_details cd
+on cd.outcome_cohort_id = o.cohort_definition_id
+and cd.cohort_type = 'O'
+
+) temp_ts
+
+) temp_ts2;
+
+
+drop table if exists #agg_cohorts_after;
+select * into #agg_cohorts_after
+
+from
+(
+-- cases indexed at O
+select
+tno.subject_id,
+tno.outcome_start_date as cohort_start_date,
+tno.outcome_end_date as cohort_end_date,
+cd.cohort_definition_id
+from #target_outcome_tar tno
+INNER JOIN #cohort_details cd
+on cd.target_cohort_id = tno.target_cohort_id
+and cd.outcome_cohort_id = tno.outcome_cohort_id
+and cd.cohort_type = 'OnT'
+
+) temp_ts2;
+
+
+drop table if exists #agg_cohorts_between;
+select * into #agg_cohorts_between
+
+from
+(
+-- cases with T start and O start as end
+select
+tno.subject_id,
+tno.cohort_start_date,
+tno.outcome_start_date as cohort_end_date,
+cd.cohort_definition_id
+from #target_outcome_tar tno
+INNER JOIN #cohort_details cd
+on cd.target_cohort_id = tno.target_cohort_id
+and cd.outcome_cohort_id = tno.outcome_cohort_id
+and cd.cohort_type = 'TnObetween'
+
+) temp_ts2;
+
+drop table if exists #agg_cohorts_extras;
+select * into #agg_cohorts_extras
+
+from
+(
+
+select distinct * from (
+
+-- targets with restrictions
+
 select
 t.subject_id,
 t.cohort_start_date,
 t.cohort_end_date,
 cd.cohort_definition_id
-from #targets_agg_all as t
+from #targets_inclusions as t
 INNER JOIN #cohort_details cd
 on cd.target_cohort_id = t.cohort_definition_id
-and cd.cohort_type = 'allT'
+and cd.cohort_type = 'Tall'
 
 union
 
+-- outcomes with restrictions
 select
 o.subject_id,
 o.cohort_start_date,
 o.cohort_end_date,
 cd.cohort_definition_id
-from #outcomes_agg as o
+from #outcomes_washout as o
 INNER JOIN #cohort_details cd
 on cd.outcome_cohort_id = o.cohort_definition_id
-and cd.cohort_type = 'O'
-
-union
-
-select
-o.subject_id,
-o.cohort_start_date,
-o.cohort_end_date,
-cd.cohort_definition_id
-from #outcomes_agg_first as o
-INNER JOIN #cohort_details cd
-on cd.outcome_cohort_id = o.cohort_definition_id
-and cd.cohort_type = 'firstO'
+and cd.cohort_type = 'Oall'
 
 ) temp_ts